putting everything into folders
